pyproject.toml
[project]
name = "quant-news-codex"
version = "0.1.0"
description = "Automated Binance announcement watcher with manual approval dashboard and futures execution"
authors = ["User <user@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy>=2.0",
    "alembic",
    "psycopg[binary]",
    "pydantic>=2.7",
    "pydantic-settings",
    "httpx",
    "apscheduler",
    "jinja2",
    "python-binance",
    "loguru",
    "aiosmtplib",
    "python-dateutil",
    "python-multipart",
    "python-dotenv",
    "websocket-client"
]

[tool.uvicorn]
app = "app.main:app"
host = "0.0.0.0"
port = 8000

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
